requirements.txt
appdirs==1.4.3
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
Django==1.11.1
django-autoslug==1.9.3
django-ordered-model==1.4.3
django-social-share==1.1.1
django-tinymce==2.6.0
enum34==1.1.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
packaging==16.8
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
pytz==2017.2
singledispatch==3.4.0.3
six==1.10.0
wrapt==1.10.11
